chore: flattens thresholds into flags
now that we only have two thresholds it makes sense to flat those into
the flags struct.
