+++ src/daemons.rs
@@ -1,4 +1,4 @@
-use super::thresholds;
+use super::arguments;
 use super::events;
 use super::metrics;
 use super::processes;
@@ -29,7 +29,7 @@ struct SignalRecord {
 pub struct Monitor<T: processes::ProcessProvider, S: events::Sender, U: signals::Sender> {
     signal_sender: U,
     sink: S,
-    thresholds: thresholds::Checker,
+    thresholds: arguments::ThresholdsChecker,
     processes_discover: T,
     loop_interval: time::Duration,
     last_signals: Cache<i32, SignalRecord>,
@@ -45,7 +45,7 @@ impl<T: processes::ProcessProvider, S: events::Sender, U: signals::Sender> Monit
     // time we signaled a process.
     pub fn new(
         sink: S,
-        thresholds: thresholds::Checker,
+        thresholds: arguments::ThresholdsChecker,
         processes_discover: T,
         signal_sender: U,
     ) -> Self {
@@ -145,7 +145,6 @@ impl<T: processes::ProcessProvider, S: events::Sender, U: signals::Sender> Monit
                         continue;
                     }
                 };
-
 
                 // we only report metrics once the passes counter zeroes out.
                 if passes == 0 {
