+++ src/lib.rs
@@ -6,4 +6,3 @@ pub mod metrics;
 pub mod processes;
 pub mod signals;
 pub mod system;
-pub mod thresholds;
